refactor: 키워드 뷰모델을 상세탐색 뷰모델로 이동
